Fix mobile inventory pgTAP RPC argument type

+++ supabase/tests/20260820033500_mobile_inventory_e2e.test.sql
@@ -13,7 +13,7 @@ select lives_ok(
     'LIGHT_WEAR',
     true,
     true,
-    91,
+    91::smallint,
     false,
     false
   )$$,
